pyproject.toml
[tool.black]
line-length = 100
[tool.poetry]
name = "ecommercebackend"
version = "0.1.0"
description = ""
authors = ["Iago Estrela <IagoFilipe.EstrelaBarros@windriver.com>"]

[tool.poetry.dependencies]
python = "^3.9"
aiosqlite = "0.17.0"
alembic = "1.7.7"
anyio = "3.6.1"
asgiref = "3.5.1"
asyncpg = "0.25.0"
bcrypt = "3.2.2"
black = "22.3.0"
CacheControl = "0.12.11"
cachy = "0.3.0"
certifi = "2022.5.18.1"
cffi = "1.15.0"
charset-normalizer = "2.0.12"
cleo = "0.8.1"
click = "8.1.3"
clikit = "0.6.2"
crashtest = "0.3.1"
cryptography = "37.0.2"
databases = "0.5.5"
distlib = "0.3.4"
ecdsa = "0.17.0"
fastapi = "0.78.0"
filelock = "3.7.0"
greenlet = "1.1.2"
h11 = "0.13.0"
html5lib = "1.1"
idna = "3.3"
importlib-metadata = "4.11.4"
isort = "5.10.1"
jeepney = "0.8.0"
keyring = "23.5.0"
lockfile = "0.12.2"
Mako = "1.2.0"
MarkupSafe = "2.1.1"
msgpack = "1.0.3"
mypy-extensions = "0.4.3"
ormar = "0.11.0"
packaging = "20.9"
passlib = "1.7.4"
pastel = "0.2.1"
pathspec = "0.9.0"
pexpect = "4.8.0"
pkginfo = "1.8.2"
platformdirs = "2.5.2"
poetry = "1.1.13"
poetry-core = "1.0.8"
psycopg2 = "2.9.3"
psycopg2-binary = "2.9.3"
ptyprocess = "0.7.0"
pyasn1 = "0.4.8"
pycparser = "2.21"
pydantic = "1.9.0"
pylev = "1.4.0"
pyparsing = "3.0.9"
python-jose = "3.3.0"
python-multipart = "0.0.5"
requests = "2.27.1"
requests-toolbelt = "0.9.1"
rsa = "4.8"
SecretStorage = "3.3.2"
shellingham = "1.4.0"
six = "1.16.0"
sniffio = "1.2.0"
SQLAlchemy = "1.4.31"
starlette = "0.19.1"
tomli = "2.0.1"
tomlkit = "0.10.2"
typing-extensions = "4.2.0"
urllib3 = "1.26.9"
uvicorn = "0.17.6"
virtualenv = "20.14.1"
webencodings = "0.5.1"
zipp = "3.8.0"
aioredis = "^2.0.1"
redis = "^4.3.3"
itsdangerous = "^2.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
